pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langchain-tools>=0.1.34",
    "langgraph>=0.6.7",
    "langgraph-prebuilt>=0.6.4",
    "python-dotenv>=1.1.1",
    "tenacity==8.5.0",
]
